test(installer): consolidate deferred onboarding coverage
